refactor(ecstore): split store_api modules
